Move My Projects out of Submitter.
