+++ client/src/app/app.routes.ts
@@ -5,7 +5,7 @@ import {AllProjectsComponent} from './all-projects/all-projects.component';
 import {LoginComponent} from './login/login.component';
 import {AaiCallbackComponent} from './aai-callback/aai-callback.component';
 import {ProjectFormComponent} from './submitter/project-form/project-form.component';
-import {MyProjectsComponent} from './submitter/my-projects/my-projects.component';
+import {MyProjectsComponent} from './my-projects/my-projects.component';
 import {SubmissionListComponent} from './submission-list/submission-list.component';
 import {RegistrationComponent} from './registration/registration.component';
 import {UserIsLoggedInGuard} from './shared/guards/user-is-logged-in.guard';
